CNN
Application of Convolutional neural networks on data sets. the aim is to describe human activity from smartphones.
The experiments were carried out with a group of 30 volunteers within an age bracket of 19-48 years. They performed a protocol of activities composed of six basic activities: three static postures (standing, sitting, lying) and three dynamic activities (walking, walking downstairs and walking upstairs). All the participants were wearing a smartphone (Samsung Galaxy S II) on the waist during the experiment execution. We captured 3-axial linear acceleration and 3-axial angular velocity at a constant rate of 50Hz using the embedded accelerometer and gyroscope of the device. The experiments were video-recorded to label the data manually. The obtained dataset was randomly partitioned into two sets, where 70% of the volunteers was selected for generating the training data and 30% the test data.
Modifications to this dataset: The data has been converted to matlab “.mat files” consisting of 6 attributes from accelerometer and body gyroscope. This code also merges the “walking upstairs and downstairs” to “climbing stairs”, so there are 5 classified activities.